Added health endpoint, fixed max file size bugs

+++ CDN.NET.Backend/Helpers/UploadSettings.cs
@@ -2,7 +2,7 @@ namespace CDN.NET.Backend.Helpers
 {
     public class UploadSettings
     {
-        public int MaxSize { get; set; }
+        public long MaxSize { get; set; }
         public int MaxFiles { get; set; }
         public string[] BlockedExtensions { get; set; }
         public bool Private { get; set; }
